tools/requirements.txt
pyyaml
librosa